# contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Phoenix Intelligence API
  version: 1.0.0
  description: |
    Versioned API contract between phoenix-core and phoenix-intelligence.
    No AI logic exists in phoenix-core; all reasoning is delegated here.
servers:
  - url: http://localhost:8000/api/v1

paths:
  /tests/generate:
    post:
      summary: Generate manual and automation test cases
      operationId: generateTests
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TestGenerationRequest'
      responses:
        '200':
          description: Test generation response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TestGenerationResponse'

  /locators/discover:
    post:
      summary: Discover element locators
      operationId: discoverLocators
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LocatorDiscoveryRequest'
      responses:
        '200':
          description: Locator discovery response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LocatorDiscoveryResponse'

  /failures/analyze:
    post:
      summary: Analyze test failure
      operationId: analyzeFailure
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FailureAnalysisRequest'
      responses:
        '200':
          description: Failure analysis response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FailureAnalysisResponse'

components:
  schemas:
    TestGenerationRequest:
      type: object
      required: [user_story, application_url]
      properties:
        user_story:
          type: string
        application_url:
          type: string
        acceptance_criteria:
          type: array
          items:
            type: string
        options:
          type: object
          properties:
            test_type:
              type: string
              enum: [manual, automation, both]
            risk_level:
              type: string
              enum: [smoke, regression, edge]
            output_style:
              type: string
              enum: [markdown, gherkin]
    TestGenerationResponse:
      type: object
      properties:
        manual_tests:
          type: array
          items:
            $ref: '#/components/schemas/ManualTestCase'
        automation_tests:
          type: array
          items:
            $ref: '#/components/schemas/AutomationTestCase'
        metadata:
          $ref: '#/components/schemas/ResponseMetadata'

    ManualTestCase:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        steps:
          type: array
          items:
            type: string
        expected_result:
          type: string
        risk_level:
          type: string
        tags:
          type: array
          items:
            type: string

    AutomationTestCase:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        test_steps:
          type: array
          items:
            type: string
        locators:
          type: array
          items:
            $ref: '#/components/schemas/Locator'
        script_template:
          type: string
        script_code:
          type: string
        risk_level:
          type: string
        tags:
          type: array
          items:
            type: string

    LocatorDiscoveryRequest:
      type: object
      required: [page_url, elements]
      properties:
        page_url:
          type: string
        elements:
          type: array
          items:
            type: string
        dom_snapshot:
          type: string
    LocatorDiscoveryResponse:
      type: object
      properties:
        locators:
          type: array
          items:
            $ref: '#/components/schemas/Locator'
        recommended_locator:
          $ref: '#/components/schemas/Locator'
        metadata:
          $ref: '#/components/schemas/ResponseMetadata'

    FailureAnalysisRequest:
      type: object
      required: [error_message]
      properties:
        error_message:
          type: string
        traceback:
          type: string
        test_context:
          type: object
    FailureAnalysisResponse:
      type: object
      properties:
        failure_reason:
          type: string
        suggested_fixes:
          type: array
          items:
            type: string
        metadata:
          $ref: '#/components/schemas/ResponseMetadata'

    Locator:
      type: object
      properties:
        element:
          type: string
        selector:
          type: string
        strategy:
          type: string
        confidence:
          type: number
          format: float

    ResponseMetadata:
      type: object
      properties:
        request_id:
          type: string
        version:
          type: string
        generated_at:
          type: string
          format: date-time